requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
mistralai>=0.0.7
mysql-connector-python>=8.2.0
pymemcache>=4.0.0